+++ MiniWin/ui/ui_check_box.h
@@ -84,7 +84,7 @@ typedef struct
 mw_handle_t mw_ui_check_box_add_new(int16_t x,
 		int16_t y,
 		mw_handle_t parent_handle,
-		uint32_t flags,
+		uint16_t flags,
 		mw_ui_check_box_data_t *check_box_instance_data);
 
 #ifdef __cplusplus
